requirements.txt
Flask
gunicorn
requests
python-docx
python-dotenv
tqdm
pdf2docx
